requirements.txt
Flask==1.1.2
py-cpuinfo==7.0.0
psutil==5.7.2
gunicorn